+++ src/steps/resource-manager/api-management/index.ts
@@ -22,6 +22,11 @@ import {
 } from './converters';
 import createResourceGroupResourceRelationship from '../utils/createResourceGroupResourceRelationship';
 import { STEP_RM_RESOURCES_RESOURCE_GROUPS } from '../resources';
+import {
+  createDiagnosticSettingsEntitiesAndRelationshipsForResource,
+  diagnosticSettingsEntitiesForResource,
+  diagnosticSettingsRelationshipsForResource,
+} from '../utils/createDiagnosticSettingsEntitiesAndRelationshipsForResource';
 export * from './constants';
 
 export async function fetchApiManagementServices(
@@ -41,6 +46,11 @@ export async function fetchApiManagementServices(
     await jobState.addEntity(apiManagementServiceEntity);
 
     await createResourceGroupResourceRelationship(
+      executionContext,
+      apiManagementServiceEntity,
+    );
+
+    await createDiagnosticSettingsEntitiesAndRelationshipsForResource(
       executionContext,
       apiManagementServiceEntity,
     );
@@ -84,8 +94,14 @@ export const apiManagementSteps: Step<
   {
     id: STEP_RM_API_MANAGEMENT_SERVICES,
     name: 'API Management Services',
-    entities: [ApiManagementEntities.SERVICE],
-    relationships: [ApiManagementRelationships.RESOURCE_GROUP_HAS_SERVICE],
+    entities: [
+      ApiManagementEntities.SERVICE,
+      ...diagnosticSettingsEntitiesForResource,
+    ],
+    relationships: [
+      ApiManagementRelationships.RESOURCE_GROUP_HAS_SERVICE,
+      ...diagnosticSettingsRelationshipsForResource,
+    ],
     dependsOn: [STEP_AD_ACCOUNT, STEP_RM_RESOURCES_RESOURCE_GROUPS],
     executionHandler: fetchApiManagementServices,
   },